fix: address code review findings for update command and GUI endpoints
- Extract fetchLatestVersion to src/lib/update-checker.js (shared between CLI and GUI)
- Stop installSpinner before execa stdio:inherit to prevent interleaved ANSI output
- Add concurrency guard (updateInProgress flag) to /api/update/stream; returns 409 on concurrent requests
- Move execa and readline imports to top-level in gui-server.js
- Update tests to mock fetchLatestVersion directly instead of stubbing global fetch
